requirements.txt
datasets==2.15.0
evaluate==0.4.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
peft==0.11.1
PyYAML==6.0.1
scikit_learn==1.3.2
tabulate==0.9.0
torch==2.1.1
tqdm==4.66.1
transformers==4.41.2
trl==0.8.6
gdown==5.1.0
bitsandbytes==0.43.0
einops==0.7.0
captum==0.7.0
click==8.1.7
jupyter==1.0.0
